requirements.txt
# Dashboard de Amenaza Geotecnica - Corredor Vial
# Instalar con: pip install -r requirements.txt
# O mejor, usar conda para geopandas/rasterio:
#   conda install -c conda-forge geopandas rasterio
#   pip install streamlit streamlit-folium folium pillow

# Core
streamlit>=1.31.0
streamlit-folium>=0.19.0
folium>=0.15.0

# Geospatial (preferir instalacion via conda)
geopandas>=0.14.0
rasterio>=1.3.9
shapely>=2.0.0

# Data
pandas>=2.1.0
numpy>=1.25.0

# Image processing
Pillow>=10.0.0
